requirements.txt
ipython
ollama
py5[extras]
pandas
python-fasthtml
anthropic